+++ lab-07-orm-queries/src/main/java/com/cydeo/lab07ormqueries/repository/CartItemRepository.java
@@ -8,10 +8,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface CartItemRepository extends JpaRepository<CartItem, Long> {
+public interface CartItemRepository extends JpaRepository<CartItem, BigInteger> {
 
     //Write a derived query to get count cart items
     long countBy();
