+++ lab-07-orm-queries/src/main/java/com/cydeo/lab07ormqueries/repository/CartRepository.java
@@ -1,4 +1,5 @@
 package com.cydeo.lab07ormqueries.repository;
+
 import com.cydeo.lab07ormqueries.entity.Cart;
 import com.cydeo.lab07ormqueries.entity.Customer;
 import com.cydeo.lab07ormqueries.enums.CartState;
@@ -7,10 +8,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface CartRepository extends JpaRepository<Cart, Long> {
+public interface CartRepository extends JpaRepository<Cart, BigInteger> {
 
     //Write a derived query to get all cart by specific discount type
     List<Cart> findByDiscount_DiscountType(DiscountType discountType);
@@ -27,5 +29,5 @@ public interface CartRepository extends JpaRepository<Cart, Long> {
 
     //Write a native query to get all cart by customer and cart state and discount is not null condition
     @Query(value = "SELECT * FROM cart WHERE customer_id = ?1 AND cart_state = ?2 AND discount_id IS NOT NULL", nativeQuery = true)
-    List<Cart> findByCustomerAndCartStateAndDiscountNotNull(Long customerId, CartState cartState);
+    List<Cart> findByCustomerAndCartStateAndDiscountNotNull(BigInteger customerId, CartState cartState);
 }
